Fixing parameter changes

+++ src/lib-sana/Graph/Graph.hpp
@@ -19,7 +19,7 @@ public:
     unsigned int GetNumNodes() const;
     unsigned int GetNumEdges() const;
 
-    virtual void SetNumNodes(const unsigned long long int &);
+    virtual void SetNumNodes(const unsigned int &);
 
     string getName() const;
